+++ gulpfile.js
@@ -21,6 +21,7 @@ const commonConfig = {
 const UMD_BOOTSTRAPPER = './umd-bootstrapper.js';
 const UMD_BOOTSTRAPPER_MINIFIED = './umd-bootstrapper-minified.js';
 const UMD_BOOTSTRAPPER_WITH_DEFAULT = './umd-bootstrapper-with-default.js';
+const UMD_BOOTSTRAPPER_WITH_DEFAULT_MINIFIED = './umd-bootstrapper-with-default-minified.js';
 
 // Package files into directory with version number.
 const VERSION_NUMBER = process.env.VERSION_NUMBER;
@@ -48,8 +49,47 @@ gulp.task('build-session-client', () => {
       .pipe(gulp.dest('./bin'))
 });
 
+gulp.task('build-minified-session-client', () => {
+  const taskConfig = {
+    js: SESSION_CLIENT_SRC,
+    js_output_file: 'omid-session-client-v1beta.min.js',
+    compilation_level: 'ADVANCED',
+    formatting: 'SINGLE_QUOTES',
+    output_wrapper_file: UMD_BOOTSTRAPPER_WITH_DEFAULT_MINIFIED,
+    externs: [
+      ...commonConfig.externs,
+      './src/externs/omid-session-client.js',
+      './src/externs/omid-jasmine.js',
+      './src/externs/omid-exports.js',
+    ],
+  };
+  return closureCompiler(Object.assign({}, commonConfig, taskConfig))
+      .src() // needed to force the plugin to run without gulp.src
+      .pipe(gulp.dest('./bin'))
+});
+
+gulp.task('build-modern-session-client', () => {
+  const taskConfig = {
+    js: SESSION_CLIENT_SRC,
+    js_output_file: 'omid-session-client-v1beta.modern.js',
+    output_wrapper_file: UMD_BOOTSTRAPPER_WITH_DEFAULT,
+    externs: [
+      ...commonConfig.externs,
+      './src/externs/omid-exports.js',
+      './src/externs/omid-jasmine.js',
+    ],
+    language_out: 'ECMASCRIPT_2017',
+    rewrite_polyfills: false,
+  };
+  return closureCompiler(Object.assign({}, commonConfig, taskConfig))
+      .src() // needed to force the plugin to run without gulp.src
+      .pipe(gulp.dest('./bin'))
+});
+
 const SESSION_CLIENT_ZIP_SRC = [
   './bin/omid-session-client-v1.js',
+  './bin/omid-session-client-v1beta.min.js',
+  './bin/omid-session-client-v1beta.modern.js',
   './LICENSE',
 ];
 
@@ -90,7 +130,7 @@ gulp.task('build-verification-client', () => {
 gulp.task('build-minified-verification-client', () => {
   const taskConfig = {
     js: VERIFICATION_CLIENT_SRC,
-    js_output_file: 'omid-verification-client-v1.min.js',
+    js_output_file: 'omid-verification-client-v1beta.min.js',
     compilation_level: 'ADVANCED',
     formatting: 'SINGLE_QUOTES',
     output_wrapper_file: UMD_BOOTSTRAPPER_MINIFIED,
@@ -108,8 +148,30 @@ gulp.task('build-minified-verification-client', () => {
       .pipe(gulp.dest('./bin'))
 });
 
+gulp.task('build-modern-verification-client', () => {
+  const taskConfig = {
+    js: VERIFICATION_CLIENT_SRC,
+    js_output_file: 'omid-verification-client-v1beta.modern.js',
+    output_wrapper_file: UMD_BOOTSTRAPPER,
+    dependency_mode: 'PRUNE',
+    entry_point: 'goog:omid.verificationClient.VerificationClient',
+    externs: [
+      ...commonConfig.externs,
+      './src/externs/omid-jasmine.js',
+      './src/externs/omid-exports.js',
+    ],
+    language_out: 'ECMASCRIPT_2017',
+    rewrite_polyfills: false,
+  };
+  return closureCompiler(Object.assign({}, commonConfig, taskConfig))
+      .src() // needed to force the plugin to run without gulp.src
+      .pipe(gulp.dest('./bin'))
+});
+
 const VERIFICATION_CLIENT_ZIP_SRC = [
   './bin/omid-verification-client-v1.js',
+  './bin/omid-verification-client-v1beta.min.js',
+  './bin/omid-verification-client-v1beta.modern.js',
   './LICENSE',
 ];